+++ SPYROpenGL/src/Dessin.cpp
@@ -801,15 +801,15 @@ const void Dessin::TroisiemePartieQueue(float longueurX, float longueurZ, float 
  */
 const void Dessin::Jambes(float longueurX, float longueurZ, float hauteur, bool sens, float agrandissement, float largeurDUnCoteDuCorps)
 {
-    float h = (float) sqrt(longueurX*longueurX + longueurZ*longueurZ); // hypoténuse du triangle du dessus
+    float h = sqrtf(longueurX*longueurX + longueurZ*longueurZ); // hypoténuse du triangle du dessus
     float t = agrandissement;
-    float tx = (float) sqrt(t*t/2);
+    float tx = sqrtf(t*t/2);
 
     // longueurs du triangle de base et de l'agrandit
-    float hypotenuseDeBase = (float) sqrt(2*(longueurX * longueurX));
+    float hypotenuseDeBase = sqrtf(2*(longueurX * longueurX));
     float hypotenuseAgrandie = largeurDUnCoteDuCorps;
     float coteDeBase = longueurX;
-    float coteAgrandie = (float) sqrt(pow(hypotenuseAgrandie,2)/2);
+    float coteAgrandie = sqrtf(pow(hypotenuseAgrandie,2)/2);
 
     // coefficients d'agrandissement
     float coefficientAgrandissement = hypotenuseAgrandie/hypotenuseDeBase;
@@ -817,10 +817,10 @@ const void Dessin::Jambes(float longueurX, float longueurZ, float hauteur, bool 
     float diagonaleAgrandissement = (coteDeBase*(hypotenuseAgrandie/hypotenuseDeBase))-coteDeBase;
 
     // calculs vecteurs de translation :
-    float translation = (float) (sqrt(pow(diagonaleAgrandissement,2) + pow((agrandissementHypotenuseParThales-hypotenuseDeBase)/2,2)) /1.75);
+    float translation = sqrtf(pow(diagonaleAgrandissement,2) + pow((agrandissementHypotenuseParThales-hypotenuseDeBase)/2,2)) /1.75;
 
     // calculs de rotation de la face du dessus
-    float rotationAxeVertical = (float) (sqrt(pow(largeurDUnCoteDuCorps,2) - 0.2*0.2));
+    float rotationAxeVertical = sqrtf(pow(largeurDUnCoteDuCorps,2) - 0.2*0.2);
     float rotationAxeVerticalMoitie = rotationAxeVertical/2;
     float rotationAxeHorizontal = hypotenuseAgrandie - 0.2f;
     float rotationAxeHorizontalMoitie = rotationAxeHorizontal/2;
